+++ src/platform/archive.hpp
@@ -4,6 +4,7 @@
 #include <vector>
 #include <functional>
 #include <filesystem>
+#include <cstdint>
 
 namespace platform {
 
@@ -16,7 +17,7 @@ void create_tar(const std::filesystem::path& tar_path,
 // Stream a tar archive through a callback. write_fn is called with each
 // chunk of tar data; it must return the number of bytes consumed (or < 0
 // on error). No temp file is created.
-using TarWriteFn = std::function<ssize_t(const void*, size_t)>;
+using TarWriteFn = std::function<int64_t(const void*, size_t)>;
 void create_tar_to_callback(const std::filesystem::path& base_dir,
                              const std::vector<std::string>& files,
                              TarWriteFn write_fn);
